--- a/ape_infura/__init__.py
+++ b/ape_infura/__init__.py
@@ -1,41 +1,22 @@
 from ape import plugins
 
-from .provider import Infura
-
-NETWORKS = {
-    "arbitrum": [
-        "mainnet",
-        "sepolia",
-    ],
-    "avalanche": [
-        "fuji",
-        "mainnet",
-    ],
-    "blast": [
-        "mainnet",
-        "sepolia",
-    ],
-    "ethereum": [
-        "mainnet",
-        "sepolia",
-    ],
-    # TODO: Comment out after ape-linea supports 0.7
-    # "linea": [
-    #     "mainnet",
-    # ],
-    "optimism": [
-        "mainnet",
-        "sepolia",
-    ],
-    "polygon": [
-        "mainnet",
-        "amoy",
-    ],
-}
-
 
 @plugins.register(plugins.ProviderPlugin)
 def providers():
+    from ape_infura.provider import Infura
+    from ape_infura.utils import NETWORKS
+
     for ecosystem_name in NETWORKS:
         for network_name in NETWORKS[ecosystem_name]:
             yield ecosystem_name, network_name, Infura
+
+
+def __getattr__(name: str):
+    if name == "NETWORKS":
+        from ape_infura.utils import NETWORKS
+
+        return NETWORKS
+
+    import ape_infura.provider as module
+
+    return getattr(module, name)