--- a/ape_infura/utils.py
+++ b/ape_infura/utils.py
@@ -1,0 +1,30 @@
+NETWORKS = {
+    "arbitrum": [
+        "mainnet",
+        "sepolia",
+    ],
+    "avalanche": [
+        "fuji",
+        "mainnet",
+    ],
+    "blast": [
+        "mainnet",
+        "sepolia",
+    ],
+    "ethereum": [
+        "mainnet",
+        "sepolia",
+    ],
+    # TODO: Comment out after ape-linea supports 0.7
+    # "linea": [
+    #     "mainnet",
+    # ],
+    "optimism": [
+        "mainnet",
+        "sepolia",
+    ],
+    "polygon": [
+        "mainnet",
+        "amoy",
+    ],
+}