### What I did

Before:

```
In [1]: %time import ape_infura
CPU times: user 1.09 s, sys: 149 ms, total: 1.24 s
Wall time: 1.25 s
```

After:

```
In [1]: %time import ape_infura
CPU times: user 7.45 ms, sys: 2.55 ms, total: 10 ms
Wall time: 8.9 ms
```

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)